Change int to long long for variable n

+++ CSES Problem Set/Weird Algorithm/main.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 int main()
 {
-    int n = 0;
+    long long n = 0;
 
     cin >> n;
     while (n != 1)  {
@@ -19,6 +19,7 @@ int main()
 
     return 0;
 }
+
